Super-initial EDA + Context for Predictors
Superficial EDA containing description of variables, some assessment for relationship to response for continuous variables, and frequency analysis for categorical variables just to see what we're working with.

Added background information to better understand the data and what the explanatory variables are telling us. Will help as sanity checks when we begin building models.